Copy simple_mock stub docs and types
